# assets/api.yaml
openapi: 3.0.0
paths:
...
  /userRoles:
    get:
      tags:
        - UserRoles
      summary: 'Получение списка записей'
      description: ''
      operationId: getUserRoles
      parameters:
        -
          name: search
          in: query
          description: 'поиск по определенному столбцу'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'сортировка по определенному столбцу'
          schema:
            type: object
        -
          name: page
          in: query
          description: Страница
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'Элементов на странице'
          schema:
            type: int32
      responses:
        '200':
          description: 'успешная операция'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserRoles' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/UserRoles' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/UserRoles' } }
                type: object
        '404':
          description: 'UserRoles не найден'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - UserRoles
      summary: 'Добавить запись'
      operationId: addUserRoles
      requestBody:
        $ref: '#/components/requestBodies/UserRoles'
      responses:
        '201':
          description: 'Created UserRoles'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoles'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserRoles'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/userRoles/{id}':
    get:
      tags:
        - UserRoles
      summary: 'Возвращает одну запись'
      description: 'Returns a single UserRoles'
      operationId: getUserRolesById
      parameters:
        -
          name: id
          in: path
          description: 'ID of UserRoles to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoles'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserRoles'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'UserRoles not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - UserRoles
      summary: 'Обновить запись'
      operationId: updateUserRoles
      parameters:
        -
          name: id
          in: path
          description: 'UserRoles id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/UserRoles'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'UserRoles not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - UserRoles
      summary: 'Удалить запись'
      operationId: deleteUserRoles
      parameters:
        -
          name: id
          in: path
          description: 'UserRoles id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    UserRoles:
      title: UserRoles
      description: UserRoles
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 11
          nullable: false
        id_subdivisions:
          title: id_subdivisions
          description: id_subdivisions
          type: integer
          format: '-'
          maxLength: 10
          nullable: false
        title:
          title: title
          description: title
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        status:
          title: status
          description: status
          type: integer
          format: '-'
          maxLength: 1
          nullable: false
      type: object
  requestBodies:
...
    UserRoles:
      description: 'UserRoles object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/UserRoles'
        application/xml:
          schema:
            $ref: '#/components/schemas/UserRoles'
        application/json:
          schema:
            $ref: '#/components/schemas/UserRoles'
...